Addressed review comments, minor changes
