+++ bee-ternary/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bee-ternary"
-version = "0.1.0"
+version = "0.1.0-alpha"
 authors = ["IOTA Stiftung"]
 edition = "2018"
 description = "Ergonomic ternary manipulation utilities"
