+++ bee-ternary/tests/common.rs
@@ -9,6 +9,7 @@
 // an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and limitations under the License.
 
+// Workaround for cargo/rustc unused warning bug
 #![allow(dead_code)]
 
 use bee_ternary::*;
